[Refactor] constants.ts 리팩터링 (#44)
* refactor: 기존 contstants.ts를 announcement feature로 이동

* refactor: 기존 constants.ts 경로를 리팩터링한 경로로 변경

* refactor: import 경로에 "@" 사용으로 통일

* fix: @/.../.module.scss 패턴 지정
