+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.scss';
-import App from './App';
+import App from '@/App';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
